Bug Fix: Bad Inline Functions
In this commit, I introduce a workaround for an odd weird bug when
using inline functions. It seems that GCC is so aggressive when
optimizing inline functions that it pass bad pointers to these
functions. To solve this, I basically made all inline functions
normal functions. Performance will not be greatly hit do to calls to
utility functions, so I think that's fine.